Update _config.yaml
Shifting the theme over to Minima V3 with remote theme. I'm not tying this to a specific commit there so things may get weird. But that's fine, I just want to see if the site will build at all doing this.

I added in all the stuff from V3 that was breaking Minima V2.5.1 from a previous commit.